refactor(form-builder): rename PTE css files to module.css
This will allow them to work in the Workshop
