+++ packages/@sanity/form-builder/src/inputs/PortableText/Editor.tsx
@@ -19,7 +19,7 @@ import {FOCUS_TERMINATOR} from '@sanity/util/paths'
 import {ScrollContainer} from '@sanity/base/components'
 import React, {useMemo, useCallback, useEffect, useState} from 'react'
 import PatchEvent from '../../PatchEvent'
-import styles from './PortableTextInput.css'
+import styles from './PortableTextInput.module.css'
 import Toolbar from './Toolbar/Toolbar'
 import {ExpandCollapseButton} from './expandCollapseButton'
 import BlockExtrasOverlay from './BlockExtrasOverlay'
