+++ packages/@sanity/form-builder/src/inputs/PortableText/Markers.tsx
@@ -6,7 +6,7 @@ import ValidationStatus from 'part:@sanity/components/validation/status'
 import CustomMarkers from 'part:@sanity/form-builder/input/block-editor/block-markers-custom-default'
 import {Path, Marker, isValidationMarker} from '@sanity/types'
 import {RenderCustomMarkers} from './types'
-import styles from './Markers.css'
+import styles from './Markers.module.css'
 
 type Props = {
   markers: Marker[]
