+++ packages/@sanity/form-builder/src/inputs/PortableText/Text/Annotation.tsx
@@ -7,7 +7,7 @@ import {FOCUS_TERMINATOR} from '@sanity/util/paths'
 import {Path, Marker, isValidationErrorMarker} from '@sanity/types'
 import {PatchEvent} from '../../../PatchEvent'
 
-import styles from './Annotation.css'
+import styles from './Annotation.module.css'
 
 type Props = {
   value: PortableTextChild
