+++ packages/@sanity/form-builder/src/inputs/PortableText/Text/Decorator.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import styles from './Decorator.css'
+import styles from './Decorator.module.css'
 
 type Props = {
   mark: string
