+++ packages/@sanity/form-builder/src/inputs/PortableText/Toolbar/CustomIcon.module.css
